chore: moves Ramps, Multisrp, Swap and Assets specs to tests
