+++ tests/regression/assets/multichain/asset-list.spec.ts
@@ -1,16 +1,16 @@
-import { RegressionAssets } from '../../../tags';
-import WalletView from '../../../pages/wallet/WalletView';
-import FixtureBuilder from '../../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../../tests/framework/fixtures/FixtureHelper';
-import { loginToApp } from '../../../viewHelper';
-import Assertions from '../../../../tests/framework/Assertions';
-import TokenOverview from '../../../pages/wallet/TokenOverview';
-import NetworkManager from '../../../pages/wallet/NetworkManager';
-import { setupRemoteFeatureFlagsMock } from '../../../../tests/api-mocking/helpers/remoteFeatureFlagsHelper';
+import { RegressionAssets } from '../../../../e2e/tags';
+import WalletView from '../../../../e2e/pages/wallet/WalletView';
+import FixtureBuilder from '../../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../../framework/fixtures/FixtureHelper';
+import { loginToApp } from '../../../../e2e/viewHelper';
+import Assertions from '../../../framework/Assertions';
+import TokenOverview from '../../../../e2e/pages/wallet/TokenOverview';
+import NetworkManager from '../../../../e2e/pages/wallet/NetworkManager';
+import { setupRemoteFeatureFlagsMock } from '../../../api-mocking/helpers/remoteFeatureFlagsHelper';
 import {
   remoteFeatureFlagTronAccounts,
   remoteFeatureMultichainAccountsAccountDetailsV2,
-} from '../../../../tests/api-mocking/mock-responses/feature-flags-mocks';
+} from '../../../api-mocking/mock-responses/feature-flags-mocks.ts';
 
 const ETHEREUM_NAME = 'Ethereum';
 const AVAX_NAME = 'AVAX';
