+++ tests/smoke/ramps/offramp-token-amount.spec.ts
@@ -1,18 +1,15 @@
-import { loginToApp } from '../../viewHelper';
-import WalletView from '../../pages/wallet/WalletView';
-import FundActionMenu from '../../pages/UI/FundActionMenu';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import { SmokeRamps } from '../../tags';
-import { CustomNetworks } from '../../../tests/resources/networks.e2e';
-import BuildQuoteView from '../../pages/Ramps/BuildQuoteView';
-import Assertions from '../../../tests/framework/Assertions';
-import {
-  RampsRegions,
-  RampsRegionsEnum,
-} from '../../../tests/framework/Constants';
+import { loginToApp } from '../../../e2e/viewHelper.ts';
+import WalletView from '../../../e2e/pages/wallet/WalletView.ts';
+import FundActionMenu from '../../../e2e/pages/UI/FundActionMenu.ts';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder.ts';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper.ts';
+import { SmokeRamps } from '../../../e2e/tags';
+import { CustomNetworks } from '../../resources/networks.e2e';
+import BuildQuoteView from '../../../e2e/pages/Ramps/BuildQuoteView.ts';
+import Assertions from '../../framework/Assertions.ts';
+import { RampsRegions, RampsRegionsEnum } from '../../framework/Constants.ts';
 import { Mockttp } from 'mockttp';
-import { setupRegionAwareOnRampMocks } from '../../../tests/api-mocking/mock-responses/ramps/ramps-region-aware-mock-setup';
+import { setupRegionAwareOnRampMocks } from '../../api-mocking/mock-responses/ramps/ramps-region-aware-mock-setup.ts';
 
 describe(SmokeRamps('Off-ramp token amounts'), () => {
   beforeEach(async () => {
